Release version v2.0.0
- Removed deprecated methods and fields
- Switched to `main` branch
- Update Go version
