+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/AmirSoleimani/VoucherCodeGenerator
+
+go 1.22.2
