+++ utils.go
@@ -2,21 +2,10 @@ package vcgen
 
 import (
 	"math"
-	"math/rand"
 	"strings"
 )
 
-//return random int in the range min...max
-func randomInt(min, max int) int {
-	return min + rand.Intn(1+max-min)
-}
-
-//return random char string from charset
-func randomChar(cs []byte) string {
-	return string(cs[randomInt(0, len(cs)-1)])
-}
-
-//repeat string with one str (#)
+// repeat string with one str (#)
 func repeatStr(count uint16, str string) string {
 	return strings.Repeat(str, int(count))
 }
