--- vcgen/go.mod
@@ -1,3 +1,0 @@
-module github.com/AmirSoleimani/VoucherCodeGenerator/vcgen
-
-go 1.18
